Move errors to package, move storage to pkg
