+++ pkg/apperrors/apperrors.go
@@ -11,4 +11,10 @@ var (
 	ErrNoIDPResourceForPath            = errors.New("could not find resource matching path")
 	ErrResourceIDNotPresent            = errors.New("resource id not present in token permissions")
 	ErrNoAuthzFound                    = errors.New("no authz found")
+	ErrSessionNotFound                 = errors.New("authentication session not found")
+	ErrNoSessionStateFound             = errors.New("no session state found")
+	ErrZeroLengthToken                 = errors.New("token has zero length")
+	ErrInvalidSession                  = errors.New("invalid session identifier")
+	ErrRefreshTokenExpired             = errors.New("the refresh token has expired")
+	ErrDecryption                      = errors.New("failed to decrypt token")
 )
