+++ pkg/storage/stores_test.go
@@ -1,0 +1,37 @@
+//go:build !e2e
+// +build !e2e
+
+/*
+Copyright 2015 All rights reserved.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package storage
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestCreateStorageRedis(t *testing.T) {
+	store, err := CreateStorage("redis://127.0.0.1")
+	assert.NotNil(t, store)
+	assert.NoError(t, err)
+}
+
+func TestCreateStorageFail(t *testing.T) {
+	store, err := CreateStorage("not_there:///tmp/bolt")
+	assert.Nil(t, store)
+	assert.Error(t, err)
+}
